feat: deploy yapo scraper with actions
